Minor bug

+++ floss/language/go/coverage.py
@@ -62,7 +62,7 @@ def main():
     get_extract_stats(pe, static_strings, go_strings, args.min_length)
 
 
-def get_extract_stats(pe, all_ss_strings: List[StaticString], go_strings, min_len) -> int:
+def get_extract_stats(pe, all_ss_strings: List[StaticString], go_strings, min_len) -> float:
     all_strings = list()
     # these are ascii, extract these utf-8 to get fewer chunks (ascii may split on two-byte characters, for example)
     for ss in all_ss_strings:
